Use Yandex apt mirror, in-Docker downloads, and Docker Compose
- Dockerfile: switch apt to mirror.yandex.ru (bionic), download JDK and Firefox inside image with wget --no-check-certificate
- run.sh: remove host-side JDK download and MD5 check; simplify container/image cleanup
- Add docker-compose.yml for service start (VNC on host port 1000)
- Add .dockerignore for jdk tarball

Made-with: Cursor

+++ run.sh
@@ -1,33 +1,10 @@
 #!/bin/bash
 
-docker stop java-applet-container && docker rm java-applet-container
+docker stop java-applet-container 2>/dev/null && docker rm java-applet-container
 
-	if [ $? == 0 ]
-	then
+if [ $? -eq 0 ]; then
 	image=$(docker image ls | grep java-applet-vnc | awk '{print $3}')
-	docker rmi $image
-	fi
-
-	if [ ! -f jdk-8u202-linux-x64.tar.gz ]
-	then
-	curl -s -O https://mirrors.huaweicloud.com/java/jdk/8u202-b08/jdk-8u202-linux-x64.tar.gz
-
-		if [ $? != 0 ]
-		then
-		echo "Download error. Make sure you download jdk-8u202-linux-x64.tar.gz"
-		exit 1	
-		fi 
-	fi
-
-	if [ -f jdk-8u202-linux-x64.tar.gz ]
-	then
-	md5=$(md5sum jdk-8u202-linux-x64.tar.gz | awk {'print $1'})
-	
-		if [ $md5 != "0029351f7a946f6c05b582100c7d45b7" ]
-		then
-		echo "File error. Make sure you download jdk-8u202-linux-x64.tar.gz"
-		exit 1
-		fi
-	fi
+	[ -n "$image" ] && docker rmi $image
+fi
 
 docker build --platform linux/amd64 -t java-applet-vnc . && docker run --platform linux/amd64 -p 5901:5901 --name java-applet-container java-applet-vnc
